@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.4.0",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "^12.2.0",
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
     "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "~13.4.0",
+    "@angular/forms": "^14.3.0",
     "@angular/material": "^12.2.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@badeball/cypress-cucumber-preprocessor": "^15.1.5",
     "@ngrx/effects": "^12.5.1",
     "@ngrx/store": "^12.5.1",
@@ -32,9 +32,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler-cli": "^13.0.0",
+    "@angular-devkit/build-angular": "^14.2.12",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "^14.3.0",
     "@bahmutov/cypress-esbuild-preprocessor": "^2.2.0",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
