@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { FormGroup } from '@angular/forms';
+import { UntypedFormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-subform1',
@@ -8,7 +8,7 @@ import { FormGroup } from '@angular/forms';
 })
 export class Subform1Component implements OnInit {
 
-  @Input() form!: FormGroup;
+  @Input() form!: UntypedFormGroup;
   @Input() car: any;
   @Input() carIndex!: number;
 
